requirements.txt
django==2.2
mysqlclient
requests
django-pure-pagination
pillow